feat: implement TASK-010 shared storage and atomic transfers

+++ Source/Hearthward/Tests/InventoryTests.cpp
@@ -9,10 +9,10 @@ bool FInventoryCapacityTest::RunTest(const FString& Parameters)
     FHearthwardInventoryState Inventory;
     for (const auto& Item : HearthwardBasicItems())
         TestTrue(TEXT("Known item accepted"), Inventory.Add(Item.Id, 1) == EHearthwardInventoryResult::Success);
-    TestEqual(TEXT("Five-item weight is exactly 4.55"), Inventory.GetWeightHundredths(), 455);
+    TestEqual(TEXT("Five-item weight is exactly 4.55"), Inventory.GetWeightHundredths(), int64(455));
     FHearthwardInventoryState Arrows;
     for (int32 N = 0; N < 2000; ++N) Arrows.Add(TEXT("arrow"), 1);
-    TestEqual(TEXT("2000 arrows exactly fill the bag"), Arrows.GetWeightHundredths(), 10000);
+    TestEqual(TEXT("2000 arrows exactly fill the bag"), Arrows.GetWeightHundredths(), int64(10000));
     TestTrue(TEXT("Extra arrow rejected"), Arrows.Add(TEXT("arrow"), 1) == EHearthwardInventoryResult::CapacityExceeded);
     TestEqual(TEXT("Failure preserves count"), Arrows.GetCount(TEXT("arrow")), 2000);
     TestTrue(TEXT("Remove accepted"), Arrows.Remove(TEXT("arrow"), 1000) == EHearthwardInventoryResult::Success);
@@ -37,13 +37,13 @@ bool FInventoryFailureTest::RunTest(const FString& Parameters)
     TestTrue(TEXT("Insufficient quantity"), Inventory.Remove(TEXT("wood"), 21) == EHearthwardInventoryResult::InsufficientItems);
     TestTrue(TEXT("Unknown removal"), Inventory.Remove(TEXT("invalid"), 1) == EHearthwardInventoryResult::UnknownItem);
     TestEqual(TEXT("All failures preserve original quantity"), Inventory.GetCount(TEXT("wood")), 20);
-    TestEqual(TEXT("All failures preserve original weight"), Inventory.GetWeightHundredths(), 2000);
+    TestEqual(TEXT("All failures preserve original weight"), Inventory.GetWeightHundredths(), int64(2000));
     Inventory.Add(TEXT("ore"), 40);
     TestTrue(TEXT("Full load speed"), FMath::IsNearlyEqual(Inventory.GetMoveSpeedMultiplier(), 0.9f));
     TestTrue(TEXT("Full load stamina cost"), FMath::IsNearlyEqual(Inventory.GetStaminaCostMultiplier(), 1.1f));
     Inventory.Remove(TEXT("wood"), 20);
     Inventory.Remove(TEXT("ore"), 40);
-    TestEqual(TEXT("Empty weight restored"), Inventory.GetWeightHundredths(), 0);
+    TestEqual(TEXT("Empty weight restored"), Inventory.GetWeightHundredths(), int64(0));
     TestEqual(TEXT("Empty speed restored"), Inventory.GetMoveSpeedMultiplier(), 1.0f);
     return true;
 }
